Rename _annotated.py to _filtered.py
